All: reduce warnings
1) No semicolon at inline functions
2) Start using size_t instead of int, where posible and document where
it also needs to be changed.

+++ src/textfield.h
@@ -48,6 +48,6 @@ struct TextField : Widget {
 
     Timer timer;
     bool cursor_on = false;
-    int cursor_position = 0;
-    int display_from = 0;
+    size_t cursor_position = 0;
+    size_t display_from = 0;
 };
